:heavy_minus_sign: Remove styles folder
