+++ src/components/windowControls.jsx
@@ -5,7 +5,7 @@ import { WindowsClose }    from './close.windows'
 import { WindowsMaximize } from './maximize.windows'
 import { WindowsMinimize } from './minimize.windows'
 
-import { grey }            from '../styles/colors'
+import { grey }            from '../defaults/colors'
 
 export class WindowControls extends Component {
   getStyles(platform) {
