+++ src/defaults/colors.js
@@ -1,281 +1,310 @@
-import { mergeArrays } from '../utils/arrays'
+export const red = {
+  50:     '#FFEBEE',
+  100:    '#FFCDD2',
+  200:    '#EF9A9A',
+  300:    '#E57373',
+  400:    '#EF5350',
+  500:    '#F44336',
+  600:    '#E53935',
+  700:    '#D32F2F',
+  800:    '#C62828',
+  900:    '#B71C1C',
+  'A100': '#FF8A80',
+  'A200': '#FF5252',
+  'A400': '#FF1744',
+  'A700': '#D50000'
+}
 
-export const smallPalette = [
-  'rgba(000, 000, 000, 1.0)', // black
-  'rgba(229, 075, 075, 1.0)', // red
-  'rgba(158, 206, 088, 1.0)', // green
-  'rgba(250, 237, 112, 1.0)', // yellow
-  'rgba(057, 111, 226, 1.0)', // blue
-  'rgba(187, 128, 179, 1.0)', // magenta
-  'rgba(045, 218, 253, 1.0)', // cyan
-  'rgba(208, 208, 208, 1.0)', // white
-  'rgba(255, 255, 255, 0.2)', // bright black
-  'rgba(255, 083, 112, 1.0)', // bright red
-  'rgba(195, 232, 141, 1.0)', // bright green
-  'rgba(255, 203, 107, 1.0)', // bright yellow
-  'rgba(130, 170, 255, 1.0)', // bright blue
-  'rgba(199, 146, 234, 1.0)', // bright magenta
-  'rgba(137, 221, 255, 1.0)', // bright cyan
-  'rgba(255, 255, 255, 1.0)'  // bright white
-]
+export const pink = {
+  50:     '#FCE4EC',
+  100:    '#F8BBD0',
+  200:    '#F48FB1',
+  300:    '#F06292',
+  400:    '#EC407A',
+  500:    '#E91E63',
+  600:    '#D81B60',
+  700:    '#C2185B',
+  800:    '#AD1457',
+  900:    '#880E4F',
+  'A100': '#FF80AB',
+  'A200': '#FF4081',
+  'A400': '#F50057',
+  'A700': '#C51162'
+}
 
-export const bigPalette = [
-  '#000000',
-  '#800000',
-  '#008000',
-  '#808000',
-  '#000080',
-  '#800080',
-  '#008080',
-  '#c0c0c0',
-  '#808080',
-  '#ff0000',
-  '#00ff00',
-  '#ffff00',
-  '#0000ff',
-  '#ff00ff',
-  '#00ffff',
-  '#ffffff',
-  '#000000',
-  '#00005f',
-  '#000087',
-  '#0000af',
-  '#0000d7',
-  '#0000ff',
-  '#005f00',
-  '#005f5f',
-  '#005f87',
-  '#005faf',
-  '#005fd7',
-  '#005fff',
-  '#008700',
-  '#00875f',
-  '#008787',
-  '#0087af',
-  '#0087d7',
-  '#0087ff',
-  '#00af00',
-  '#00af5f',
-  '#00af87',
-  '#00afaf',
-  '#00afd7',
-  '#00afff',
-  '#00d700',
-  '#00d75f',
-  '#00d787',
-  '#00d7af',
-  '#00d7d7',
-  '#00d7ff',
-  '#00ff00',
-  '#00ff5f',
-  '#00ff87',
-  '#00ffaf',
-  '#00ffd7',
-  '#00ffff',
-  '#5f0000',
-  '#5f005f',
-  '#5f0087',
-  '#5f00af',
-  '#5f00d7',
-  '#5f00ff',
-  '#5f5f00',
-  '#5f5f5f',
-  '#5f5f87',
-  '#5f5faf',
-  '#5f5fd7',
-  '#5f5fff',
-  '#5f8700',
-  '#5f875f',
-  '#5f8787',
-  '#5f87af',
-  '#5f87d7',
-  '#5f87ff',
-  '#5faf00',
-  '#5faf5f',
-  '#5faf87',
-  '#5fafaf',
-  '#5fafd7',
-  '#5fafff',
-  '#5fd700',
-  '#5fd75f',
-  '#5fd787',
-  '#5fd7af',
-  '#5fd7d7',
-  '#5fd7ff',
-  '#5fff00',
-  '#5fff5f',
-  '#5fff87',
-  '#5fffaf',
-  '#5fffd7',
-  '#5fffff',
-  '#870000',
-  '#87005f',
-  '#870087',
-  '#8700af',
-  '#8700d7',
-  '#8700ff',
-  '#875f00',
-  '#875f5f',
-  '#875f87',
-  '#875faf',
-  '#875fd7',
-  '#875fff',
-  '#878700',
-  '#87875f',
-  '#878787',
-  '#8787af',
-  '#8787d7',
-  '#8787ff',
-  '#87af00',
-  '#87af5f',
-  '#87af87',
-  '#87afaf',
-  '#87afd7',
-  '#87afff',
-  '#87d700',
-  '#87d75f',
-  '#87d787',
-  '#87d7af',
-  '#87d7d7',
-  '#87d7ff',
-  '#87ff00',
-  '#87ff5f',
-  '#87ff87',
-  '#87ffaf',
-  '#87ffd7',
-  '#87ffff',
-  '#af0000',
-  '#af005f',
-  '#af0087',
-  '#af00af',
-  '#af00d7',
-  '#af00ff',
-  '#af5f00',
-  '#af5f5f',
-  '#af5f87',
-  '#af5faf',
-  '#af5fd7',
-  '#af5fff',
-  '#af8700',
-  '#af875f',
-  '#af8787',
-  '#af87af',
-  '#af87d7',
-  '#af87ff',
-  '#afaf00',
-  '#afaf5f',
-  '#afaf87',
-  '#afafaf',
-  '#afafd7',
-  '#afafff',
-  '#afd700',
-  '#afd75f',
-  '#afd787',
-  '#afd7af',
-  '#afd7d7',
-  '#afd7ff',
-  '#afff00',
-  '#afff5f',
-  '#afff87',
-  '#afffaf',
-  '#afffd7',
-  '#afffff',
-  '#d70000',
-  '#d7005f',
-  '#d70087',
-  '#d700af',
-  '#d700d7',
-  '#d700ff',
-  '#d75f00',
-  '#d75f5f',
-  '#d75f87',
-  '#d75faf',
-  '#d75fd7',
-  '#d75fff',
-  '#d78700',
-  '#d7875f',
-  '#d78787',
-  '#d787af',
-  '#d787d7',
-  '#d787ff',
-  '#d7af00',
-  '#d7af5f',
-  '#d7af87',
-  '#d7afaf',
-  '#d7afd7',
-  '#d7afff',
-  '#d7d700',
-  '#d7d75f',
-  '#d7d787',
-  '#d7d7af',
-  '#d7d7d7',
-  '#d7d7ff',
-  '#d7ff00',
-  '#d7ff5f',
-  '#d7ff87',
-  '#d7ffaf',
-  '#d7ffd7',
-  '#d7ffff',
-  '#ff0000',
-  '#ff005f',
-  '#ff0087',
-  '#ff00af',
-  '#ff00d7',
-  '#ff00ff',
-  '#ff5f00',
-  '#ff5f5f',
-  '#ff5f87',
-  '#ff5faf',
-  '#ff5fd7',
-  '#ff5fff',
-  '#ff8700',
-  '#ff875f',
-  '#ff8787',
-  '#ff87af',
-  '#ff87d7',
-  '#ff87ff',
-  '#ffaf00',
-  '#ffaf5f',
-  '#ffaf87',
-  '#ffafaf',
-  '#ffafd7',
-  '#ffafff',
-  '#ffd700',
-  '#ffd75f',
-  '#ffd787',
-  '#ffd7af',
-  '#ffd7d7',
-  '#ffd7ff',
-  '#ffff00',
-  '#ffff5f',
-  '#ffff87',
-  '#ffffaf',
-  '#ffffd7',
-  '#ffffff',
-  '#080808',
-  '#121212',
-  '#1c1c1c',
-  '#262626',
-  '#303030',
-  '#3a3a3a',
-  '#444444',
-  '#4e4e4e',
-  '#585858',
-  '#626262',
-  '#6c6c6c',
-  '#767676',
-  '#808080',
-  '#8a8a8a',
-  '#949494',
-  '#9e9e9e',
-  '#a8a8a8',
-  '#b2b2b2',
-  '#bcbcbc',
-  '#c6c6c6',
-  '#d0d0d0',
-  '#dadada',
-  '#e4e4e4',
-  '#eeeeee'
-]
+export const purple = {
+  50:     '#F3E5F5',
+  100:    '#E1BEE7',
+  200:    '#CE93D8',
+  300:    '#BA68C8',
+  400:    '#AB47BC',
+  500:    '#9C27B0',
+  600:    '#8E24AA',
+  700:    '#7B1FA2',
+  800:    '#6A1B9A',
+  900:    '#4A148C',
+  'A100': '#EA80FC',
+  'A200': '#E040FB',
+  'A400': '#D500F9',
+  'A700': '#AA00FF'
+}
 
-export const Colors = mergeArrays(bigPalette, smallPalette)
+export const deepPurple = {
+  50:     '#EDE7F6',
+  100:    '#D1C4E9',
+  200:    '#B39DDB',
+  300:    '#9575CD',
+  400:    '#7E57C2',
+  500:    '#673AB7',
+  600:    '#5E35B1',
+  700:    '#512DA8',
+  800:    '#4527A0',
+  900:    '#311B92',
+  'A100': '#B388FF',
+  'A200': '#7C4DFF',
+  'A400': '#651FFF',
+  'A700': '#6200EA'
+}
+
+export const indigo = {
+  50:     '#E8EAF6',
+  100:    '#C5CAE9',
+  200:    '#9FA8DA',
+  300:    '#7986CB',
+  400:    '#5C6BC0',
+  500:    '#3F51B5',
+  600:    '#3949AB',
+  700:    '#303F9F',
+  800:    '#283593',
+  900:    '#1A237E',
+  'A100': '#8C9EFF',
+  'A200': '#536DFE',
+  'A400': '#3D5AFE',
+  'A700': '#304FFE'
+}
+
+export const blue = {
+  50:     '#E3F2FD',
+  100:    '#BBDEFB',
+  200:    '#90CAF9',
+  300:    '#64B5F6',
+  400:    '#42A5F5',
+  500:    '#2196F3',
+  600:    '#1E88E5',
+  700:    '#1976D2',
+  800:    '#1565C0',
+  900:    '#0D47A1',
+  'A100': '#82B1FF',
+  'A200': '#448AFF',
+  'A400': '#2979FF',
+  'A700': '#2962FF'
+}
+
+export const lightBlue = {
+  50:     '#E1F5FE',
+  100:    '#B3E5FC',
+  200:    '#81D4FA',
+  300:    '#4FC3F7',
+  400:    '#29B6F6',
+  500:    '#03A9F4',
+  600:    '#039BE5',
+  700:    '#0288D1',
+  800:    '#0277BD',
+  900:    '#01579B',
+  'A100': '#80D8FF',
+  'A200': '#40C4FF',
+  'A400': '#00B0FF',
+  'A700': '#0091EA'
+}
+
+export const cyan = {
+  50:     '#E0F7FA',
+  100:    '#B2EBF2',
+  200:    '#80DEEA',
+  300:    '#4DD0E1',
+  400:    '#26C6DA',
+  500:    '#00BCD4',
+  600:    '#00ACC1',
+  700:    '#0097A7',
+  800:    '#00838F',
+  900:    '#006064',
+  'A100': '#84FFFF',
+  'A200': '#18FFFF',
+  'A400': '#00E5FF',
+  'A700': '#00B8D4'
+}
+
+export const team = {
+  50:     '#E0F2F1',
+  100:    '#B2DFDB',
+  200:    '#80CBC4',
+  300:    '#4DB6AC',
+  400:    '#26A69A',
+  500:    '#009688',
+  600:    '#00897B',
+  700:    '#00796B',
+  800:    '#00695C',
+  900:    '#004D40',
+  'A100': '#A7FFEB',
+  'A200': '#64FFDA',
+  'A400': '#1DE9B6',
+  'A700': '#00BFA5'
+}
+
+export const green = {
+  50:     '#E8F5E9',
+  100:    '#C8E6C9',
+  200:    '#A5D6A7',
+  300:    '#81C784',
+  400:    '#66BB6A',
+  500:    '#4CAF50',
+  600:    '#43A047',
+  700:    '#388E3C',
+  800:    '#2E7D32',
+  900:    '#1B5E20',
+  'A100': '#B9F6CA',
+  'A200': '#69F0AE',
+  'A400': '#00E676',
+  'A700': '#00C853'
+}
+
+export const lightGreen = {
+  50:     '#F1F8E9',
+  100:    '#DCEDC8',
+  200:    '#C5E1A5',
+  300:    '#AED581',
+  400:    '#9CCC65',
+  500:    '#8BC34A',
+  600:    '#7CB342',
+  700:    '#689F38',
+  800:    '#558B2F',
+  900:    '#33691E',
+  'A100': '#CCFF90',
+  'A200': '#B2FF59',
+  'A400': '#76FF03',
+  'A700': '#64DD17'
+}
+
+export const lime = {
+  50:     '#F9FBE7',
+  100:    '#F0F4C3',
+  200:    '#E6EE9C',
+  300:    '#DCE775',
+  400:    '#D4E157',
+  500:    '#CDDC39',
+  600:    '#C0CA33',
+  700:    '#AFB42B',
+  800:    '#9E9D24',
+  900:    '#827717',
+  'A100': '#F4FF81',
+  'A200': '#EEFF41',
+  'A400': '#C6FF00',
+  'A700': '#AEEA00'
+}
+
+export const yellow = {
+  50:     '#FFFDE7',
+  100:    '#FFF9C4',
+  200:    '#FFF59D',
+  300:    '#FFF176',
+  400:    '#FFEE58',
+  500:    '#FFEB3B',
+  600:    '#FDD835',
+  700:    '#FBC02D',
+  800:    '#F9A825',
+  900:    '#F57F17',
+  'A100': '#FFFF8D',
+  'A200': '#FFFF00',
+  'A400': '#FFEA00',
+  'A700': '#FFD600'
+}
+
+export const amber = {
+  50:     '#FFF8E1',
+  100:    '#FFECB3',
+  200:    '#FFE082',
+  300:    '#FFD54F',
+  400:    '#FFCA28',
+  500:    '#FFC107',
+  600:    '#FFB300',
+  700:    '#FFA000',
+  800:    '#FF8F00',
+  900:    '#FF6F00',
+  'A100': '#FFE57F',
+  'A200': '#FFD740',
+  'A400': '#FFC400',
+  'A700': '#FFAB00'
+}
+
+export const orange = {
+  50:  '#FFF3E0',
+  100: '#FFE0B2',
+  200: '#FFCC80',
+  300: '#FFB74D',
+  400: '#FFA726',
+  500: '#FF9800',
+  600: '#FB8C00',
+  700: '#F57C00',
+  800: '#EF6C00',
+  900: '#E65100'
+}
+
+export const deepOrange = {
+  50:     '#FBE9E7',
+  100:    '#FFCCBC',
+  200:    '#FFAB91',
+  300:    '#FF8A65',
+  400:    '#FF7043',
+  500:    '#FF5722',
+  600:    '#F4511E',
+  700:    '#E64A19',
+  800:    '#D84315',
+  900:    '#BF360C',
+  'A100': '#FF9E80',
+  'A200': '#FF6E40',
+  'A400': '#FF3D00',
+  'A700': '#DD2C00'
+}
+
+export const brown = {
+  50:  '#EFEBE9',
+  100: '#D7CCC8',
+  200: '#BCAAA4',
+  300: '#A1887F',
+  400: '#8D6E63',
+  500: '#795548',
+  600: '#6D4C41',
+  700: '#5D4037',
+  800: '#4E342E',
+  900: '#3E2723'
+}
+
+export const grey = {
+  50:  '#FAFAFA',
+  100: '#F5F5F5',
+  200: '#EEEEEE',
+  300: '#E0E0E0',
+  400: '#BDBDBD',
+  500: '#9E9E9E',
+  600: '#757575',
+  700: '#616161',
+  800: '#424242',
+  900: '#212121'
+}
+
+export const blueGray = {
+  50:  '#ECEFF1',
+  100: '#CFD8DC',
+  200: '#B0BEC5',
+  300: '#90A4AE',
+  400: '#78909C',
+  500: '#607D8B',
+  600: '#546E7A',
+  700: '#455A64',
+  800: '#37474F',
+  900: '#263238'
+}
+
+export const black = '#000000'
+
+export const white = '#FFFFFF'
