--- src/styles/absoluteFill.js
@@ -1,7 +1,0 @@
-export default {
-  position: 'absolute',
-  top: 0,
-  bottom: 0,
-  left: 0,
-  right: 0
-}
