--- src/styles/fixedFill.js
@@ -1,7 +1,0 @@
-export default {
-  position: 'fixed',
-  top: 0,
-  bottom: 0,
-  left: 0,
-  right: 0
-}
